requirements.txt
# Core
numpy
pandas
scikit-learn
tensorflow

# For serial communication
pyserial

# Optional (only if you use Jupyter)
jupyter
notebook
